Improve packet handling, ignore status effects
Use the 0x0C8 packet to track zoning and for updating party data when
party members are invited or kicked. This replaces the 0x0DD packet
which was sent once for each alliance member, which caused some
slowdown.

The 0x0DF packet handling has also been improved, when this packet is
now received we only update the relevant party member whose data has
been updated.

For the moment the get_status_string function is far too slow to use
even when querying status effects by id, it is even worse when using
names. For now updates to status effects will be ignored.