Fix party data updates bug
The party lib reads data from memory, but packets are processed by
addons before the client gets a chance to see them and update memory so
if you only check the values when the packets arrive, it will appear
as if the values in memory are one packet old. This was fixed by adding
a few milliseconds of delay after receiving a packet so memory has a
chance to be updated correctly when getting party data.